Outline sketch of GDARunner, UDCRunner
