+++ conftest.py
@@ -19,5 +19,5 @@ def pytest_addoption(parser):
 @pytest.fixture(scope="session", autouse=True)
 def default_session_fixture() -> Iterator[None]:
     print("Patching bluesky 0MQ Publisher in __main__ for the whole session")
-    with patch("mx_bluesky.hyperion.__main__.Publisher"):
+    with patch("mx_bluesky.hyperion.runner.Publisher"):
         yield
