+++ src/mx_bluesky/hyperion/baton_handler.py
@@ -1,5 +1,7 @@
 from collections.abc import Sequence
 
+from blueapi.core import BlueskyContext
+from bluesky import RunEngine
 from bluesky import plan_stubs as bps
 from bluesky import preprocessors as bpp
 from dodal.devices.baton import Baton
@@ -15,6 +17,7 @@ from mx_bluesky.hyperion.external_interaction.agamemnon import (
 )
 from mx_bluesky.hyperion.parameters.components import Wait
 from mx_bluesky.hyperion.parameters.load_centre_collect import LoadCentreCollect
+from mx_bluesky.hyperion.runner import BlueskyRunner, UDCRunner
 
 HYPERION_USER = "Hyperion"
 NO_USER = "None"
@@ -67,6 +70,10 @@ def main_hyperion_loop(baton: Baton, composite: LoadCentreCollectComposite):
 def move_to_default_state():
     # To be filled in in https://github.com/DiamondLightSource/mx-bluesky/issues/396
     yield from bps.null()
+
+
+def create_runner(run_engine: RunEngine, context: BlueskyContext) -> BlueskyRunner:
+    return UDCRunner(run_engine, context)
 
 
 def run_udc_when_requested(baton: Baton, composite: LoadCentreCollectComposite):
